hwmon: add driver for NZXT Grid+ V3/Smart Device (V1)
These are fan hubs that can be monitored and controlled through a
proprietary USB HID protocol.

The driver reports fan speed, current draw and supply voltage for each
channel, and supports fan control.

pwm*_enable attributes are not available as they are unusable while the
device has not completed its asynchronous initialization routine, which
can sometimes cause a fancontrol service to abort.

Reading the PWM values is not supported by the hardware, but the driver
keeps track of the values it has set, and reports those.  This is
important because user-space (and, in particular, pwmconfig/fancontrol)
expect those attributes to be RW.

LED control, which the Smart Device also supports, is not implemented
and is left to existing user-space tools; a link to one is provided in
the documentation.

The implementation is based on USB traffic analysis.  It has been tested
on x86_64, both as built-in driver and as a module.

Signed-off-by: Jonas Malaco <jonas@protocubo.io>